pyproject.toml
[project]
name = "synthetic-healthlake"
version = "0.1.0"
description = "FHIR-OMOP synthetic data pipeline reference architecture"
authors = [{name = "Robert Ford", email = "robert.ford@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
urls = {Repository = "https://github.com/robandrewford/synthetic-healthlake"}
dependencies = [
    "duckdb>=1.4.2",
    "ruff>=0.14.7",
    "pyyaml>=6.0",
    "pandas>=2.0.0",
    "faker>=20.0.0",
    "pyarrow>=14.0.0",
    "boto3>=1.34.0",
    "click>=8.1.0",
    "dbt-duckdb>=1.5.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "moto[s3]>=5.0.0",
]
dev = [
    "pre-commit>=3.5.0",
]
security = [
    "pip-audit>=2.6.0",
    "safety>=3.0.0",
    "bandit>=1.7.5",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--cov=synthetic",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv", "cdk.out"]
skips = ["B101"]  # Skip assert_used test (common in test files)

[tool.ruff]
# Ruff configuration for code quality and style
line-length = 100
target-version = "py311"

[tool.ruff.lint]
# Enable specific rule sets
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]

# Ignore specific rules
ignore = [
    "E501",   # Line too long (handled by formatter)
    "B008",   # Do not perform function calls in argument defaults
    "N805",   # First argument should be named self
]

# Exclude specific directories
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    "cdk.out",
    "node_modules",
    ".pytest_cache",
    "htmlcov",
    "history",
]

[tool.ruff.lint.per-file-ignores]
# Ignore specific rules in test files
"tests/**/*.py" = ["S101"]  # Allow assert statements in tests

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"
# Use spaces for indentation
indent-style = "space"
# Respect magic trailing comma
skip-magic-trailing-comma = false
# Auto-detect line ending
line-ending = "auto"
